+++ R/ggeffect.R
@@ -4,7 +4,6 @@
 #' @importFrom sjmisc is_empty str_contains
 #' @importFrom stats na.omit
 #' @importFrom sjlabelled as_numeric
-#' @importFrom rlang .data
 #' @importFrom insight find_predictors link_inverse print_color
 #' @export
 ggeffect <- function(model, terms, ci.lvl = .95, x.as.factor = TRUE, ...) {
@@ -33,7 +32,7 @@ ggeffect <- function(model, terms, ci.lvl = .95, x.as.factor = TRUE, ...) {
         }
       )
       no_results <- sapply(res, is.null)
-      res <- compact_list(res)
+      res <- .compact_list(res)
       if (!is.null(res) && !sjmisc::is_empty(res)) {
         names(res) <- predictors[!no_results]
         class(res) <- c("ggalleffects", class(res))
@@ -52,7 +51,7 @@ ggeffect <- function(model, terms, ci.lvl = .95, x.as.factor = TRUE, ...) {
 ggeffect_helper <- function(model, terms, ci.lvl, x.as.factor, ...) {
 
   # check terms argument
-  terms <- .check_vars(terms, model)
+  ori.terms <- terms <- .check_vars(terms, model)
   cleaned.terms <- .get_cleaned_terms(terms)
 
   # get model frame
@@ -77,6 +76,14 @@ ggeffect_helper <- function(model, terms, ci.lvl, x.as.factor, ...) {
 
   # clear argument from brackets
   terms <- .get_cleaned_terms(terms)
+
+  # check for character vectors, transform to factor
+  is_char <- sapply(terms, function(.i) is.character(fitfram[[.i]]))
+  if (any(is_char)) {
+    for (.i in terms[is_char]) {
+      fitfram[[.i]] <- as.factor(fitfram[[.i]])
+    }
+  }
 
   # fix remaining x-levels
   xl.remain <- which(!(terms %in% names(x.levels)))
@@ -116,7 +123,7 @@ ggeffect_helper <- function(model, terms, ci.lvl, x.as.factor, ...) {
   # build data frame, with raw values
   # predicted response and lower/upper ci
 
-  if (inherits(model, c("polr", "clm", "clm2", "clmm", "multinom"))) {
+  if (inherits(model, c("polr", "clm", "clm2", "clmm", "clmm2", "multinom"))) {
 
     # for categorical outcomes, we need to gather the data
     # from effects to get a single data frame
@@ -166,7 +173,7 @@ ggeffect_helper <- function(model, terms, ci.lvl, x.as.factor, ...) {
           response.level = .x,
           stringsAsFactors = FALSE
         )
-        create_eff_group(tmpl, terms, eff, sub = .x)
+        .create_eff_group(tmpl, terms, eff, sub = .x)
       })
       tmp <- do.call(rbind, l)
       fx.term <- eff[[1]]$term
@@ -181,7 +188,7 @@ ggeffect_helper <- function(model, terms, ci.lvl, x.as.factor, ...) {
           stringsAsFactors = FALSE
         )
 
-      tmp <- create_eff_group(tmp, terms, eff, sub = NULL)
+      tmp <- .create_eff_group(tmp, terms, eff, sub = NULL)
 
       # effects-package keeps the order of numeric value as they are
       # returned by "unique()", so we want to sort the data frame
@@ -219,7 +226,7 @@ ggeffect_helper <- function(model, terms, ci.lvl, x.as.factor, ...) {
   # levels for the grouping variables
 
   # for numeric values with many decimal places, we need to round
-  if (frac_length(tmp$x) > 5)
+  if (.frac_length(tmp$x) > 5)
     filter.keep <- round(tmp$x, 5) %in% round(x.levels[[1]], 5)
   else
     filter.keep <- tmp$x %in% x.levels[[1]]
@@ -268,6 +275,8 @@ ggeffect_helper <- function(model, terms, ci.lvl, x.as.factor, ...) {
   else
     xif <- ifelse(is.factor(x_v), "1", "0")
 
+  attr(mydf, "x.is.factor") <- xif
+
   # set attributes with necessary information
   mydf <-
     .set_attributes_and_class(
@@ -280,9 +289,10 @@ ggeffect_helper <- function(model, terms, ci.lvl, x.as.factor, ...) {
       legend.labels = legend.labels,
       x.axis.labels = all.labels$axis.labels,
       faminfo = faminfo,
-      x.is.factor = xif,
-      terms = cleaned.terms
+      terms = cleaned.terms,
+      ori.terms = ori.terms,
     )
+
 
   # make x numeric
   if (!x.as.factor) mydf$x <- sjlabelled::as_numeric(mydf$x, keep.labels = FALSE)
@@ -291,7 +301,7 @@ ggeffect_helper <- function(model, terms, ci.lvl, x.as.factor, ...) {
 }
 
 
-create_eff_group <- function(tmp, terms, eff, sub) {
+.create_eff_group <- function(tmp, terms, eff, sub) {
 
   if (!is.null(sub)) {
     fx <- eff[[sub]]
@@ -304,10 +314,10 @@ create_eff_group <- function(tmp, terms, eff, sub) {
     # convert to factor for proper legend
     tmp$group <- sjmisc::to_factor(1)
   } else if (length(terms) == 2) {
-    tmp$group<- sjmisc::to_factor(fx$x[[terms[2]]])
+    tmp$group <- sjmisc::to_factor(fx$x[[terms[2]]])
   } else {
-    tmp$group<- sjmisc::to_factor(fx$x[[terms[2]]])
-    tmp$facet<- sjmisc::to_factor(fx$x[[terms[3]]])
+    tmp$group <- sjmisc::to_factor(fx$x[[terms[2]]])
+    tmp$facet <- sjmisc::to_factor(fx$x[[terms[3]]])
   }
 
   tmp