+++ R/ggpredict.R
@@ -152,7 +152,8 @@
 #' @details
 #'   \subsection{Supported Models}{
 #'   Currently supported model-objects are (in alphabetical order):
-#'   \code{betabin}, \code{betareg}, \code{bglmer}, \code{blmer}, \code{brglm}, \code{brmsfit},
+#'   \code{bamlss}, \code{bayesx}, \code{betabin}, \code{betareg}, \code{bglmer},
+#'   \code{blmer}, \code{bracl}, \code{brglm}, \code{brmsfit}, \code{brmultinom},
 #'   \code{clm}, \code{clm2}, \code{clmm}, \code{coxph}, \code{gam} (package \pkg{mgcv}),
 #'   \code{Gam} (package \pkg{gam}), \code{gamlss}, \code{gamm}, \code{gamm4},
 #'   \code{gee}, \code{geeglm}, \code{glm}, \code{glm.nb}, \code{glmer},
@@ -329,6 +330,7 @@
 #'         are returned.
 #'
 #' @examples
+#' library(sjlabelled)
 #' data(efc)
 #' fit <- lm(barthtot ~ c12hour + neg_c_7 + c161sex + c172code, data = efc)
 #'
@@ -350,6 +352,7 @@
 #' # to plot ggeffects-objects, you can use the 'plot()'-function.
 #' # the following examples show how to build your ggplot by hand.
 #'
+#' \donttest{
 #' # plot predicted values, remaining covariates held constant
 #' library(ggplot2)
 #' mydf <- ggpredict(fit, terms = "c12hour")
@@ -376,7 +379,6 @@
 #'
 #' # level indication also works for factors with non-numeric levels
 #' # and in combination with numeric levels for other variables
-#' library(sjlabelled)
 #' data(efc)
 #' efc$c172code <- as_label(efc$c172code)
 #' fit <- lm(barthtot ~ c12hour + neg_c_7 + c161sex + c172code, data = efc)
@@ -412,7 +414,6 @@
 #'   )
 #'
 #' # or with ggeffects' plot-method
-#' \dontrun{
 #' plot(dat, ci = FALSE)}
 #'
 #' # use numeric values as x-column in returned data frame
@@ -431,7 +432,7 @@
 #' ggeffect(fit, terms = "e17age")
 #'
 #' @importFrom stats predict predict.glm na.omit
-#' @importFrom dplyr select mutate case_when arrange
+#' @importFrom dplyr mutate
 #' @importFrom sjmisc to_factor is_num_fac remove_empty_cols
 #' @importFrom purrr map
 #' @importFrom sjlabelled as_numeric
@@ -468,7 +469,7 @@ ggpredict <- function(model,
   # extract just the mer-part then
   if (is.gamm(model) || is.gamm4(model)) model <- model$gam
 
-  if (inherits(model, "list")) {
+  if (inherits(model, "list") && !inherits(model, "bamlss")) {
     res <- purrr::map(model, ~ggpredict_helper(
       model = .x,
       terms = terms,
@@ -617,10 +618,6 @@ ggpredict_helper <- function(model,
   # return if no predicted values have been computed
   if (is.null(fitfram)) return(NULL)
 
-
-  # init legend labels
-  legend.labels <- NULL
-
   # for survival probabilities or cumulative hazards, we need
   # the "time" variable
   if (model.class == "coxph" && type %in% c("surv", "cumhaz")) {
@@ -628,69 +625,13 @@ ggpredict_helper <- function(model,
     cleaned.terms <- c("time", cleaned.terms)
   }
 
-  # get axis titles and labels
-  all.labels <- .get_axis_titles_and_labels(
-    fitfram = ori.mf,
-    terms = terms,
-    fun = .get_model_function(model),
-    faminfo = faminfo,
-    no.transform = FALSE,
-    type = type
+  mydf <- .post_processing_predictions(
+    model = model,
+    fitfram = fitfram,
+    original.model.frame = ori.mf,
+    cleaned.terms = cleaned.terms,
+    x.as.factor = x.as.factor
   )
-
-  # check for correct terms specification
-  if (!all(terms %in% colnames(fitfram))) {
-    stop("At least one term specified in `terms` is no valid model term.", call. = FALSE)
-  }
-
-  # now select only relevant variables: the predictors on the x-axis,
-  # the predictions and the originial response vector (needed for scatter plot)
-
-  mydf <- fitfram[, stats::na.omit(match(
-    c(terms, "predicted", "conf.low", "conf.high", "response.level"),
-    colnames(fitfram)
-  ))]
-
-
-  # name and sort columns, depending on groups, facet and panel
-  mydf <- .prepare_columns(mydf, cleaned.terms)
-
-  # grouping variable may not be labelled
-  # do this here, so we convert to labelled factor later
-  mydf <- .add_labels_to_groupvariable(mydf, ori.mf, terms)
-
-  # convert grouping variable to factor, for proper legend
-  mydf <- .groupvariable_to_labelled_factor(mydf)
-
-  # check if we have legend labels
-  legend.labels <- sjlabelled::get_labels(mydf$group)
-
-  # if we had numeric variable w/o labels, these still might be numeric
-  # make sure we have factors here for our grouping and facet variables
-  if (is.numeric(mydf$group))
-    mydf$group <- sjmisc::to_factor(mydf$group)
-
-  if (obj_has_name(mydf, "facet") && is.numeric(mydf$facet)) {
-    mydf$facet <- sjmisc::to_factor(mydf$facet)
-    attr(mydf, "numeric.facet") <- TRUE
-  }
-
-
-  # remember if x is factor
-  x.is.factor <- ifelse(is.factor(mydf$x), "1", "0")
-
-  # x needs to be numeric
-  if (!x.as.factor) mydf$x <- sjlabelled::as_numeric(mydf$x)
-
-  # add standard errors
-  se <- attr(fitfram, "std.error", exact = TRUE)
-  if (is.null(se)) se <- NA
-
-  mydf <- sjmisc::add_variables(mydf, std.error = se, .after = "predicted")
-
-
-  # sort values
-  mydf <- sjmisc::remove_empty_cols(mydf[order(mydf$x, mydf$group), ])
 
   # check if outcome is log-transformed, and if so,
   # back-transform predicted values to response scale
@@ -699,28 +640,21 @@ ggpredict_helper <- function(model,
   # add raw data as well
   attr(mydf, "rawdata") <- .get_raw_data(model, ori.mf, terms)
 
-
-  # set attributes with necessary information
-  .set_attributes_and_class(
-    data = mydf,
+  .post_processing_labels(
     model = model,
-    t.title = all.labels$t.title,
-    x.title = all.labels$x.title,
-    y.title = all.labels$y.title,
-    l.title = all.labels$l.title,
-    legend.labels = legend.labels,
-    x.axis.labels = all.labels$axis.labels,
+    mydf = mydf,
+    original.model.frame = ori.mf,
+    expanded_frame = expanded_frame,
+    cleaned.terms = cleaned.terms,
+    original.terms = ori.terms,
     faminfo = faminfo,
-    x.is.factor = x.is.factor,
-    constant.values = attr(expanded_frame, "constant.values", exact = TRUE),
-    terms = cleaned.terms,
-    ori.terms = ori.terms,
+    type = type,
+    prediction.interval = attr(fitfram, "prediction.interval", exact = TRUE),
     at.list = .get_data_grid(
       model = model, mf = ori.mf, terms = ori.terms, typ.fun = typical,
       condition = condition, pretty.message = FALSE, emmeans.only = TRUE
     ),
-    n.trials = attr(expanded_frame, "n.trials", exact = TRUE),
-    prediction.interval = attr(fitfram, "prediction.interval", exact = TRUE)
+    condition = condition
   )
 }